+++ app/src/main/kotlin/org/fossify/math/helpers/converters/Base.kt
@@ -13,7 +13,10 @@ interface Converter {
             VolumeConverter,
             MassConverter,
             TemperatureConverter,
-            TimeConverter
+            TimeConverter,
+            SpeedConverter,
+            PressureConverter,
+            EnergyConverter
         )
     }